Update README.md
Generally Credit card have financial risks which potentially result in loss and affect portfolio stability. In some cases , Credit card holders may overuse the cards transaction and if they fail to pay the default due of the card , then that affects the banks income or stability .In order to avoid all these scenarios , a prediction is made by default based on history of transaction and other environmental factors involved.